test(swingset-vat): move notifier durable publish-kit upgrade integration test
